Module Superbol_preprocs

module Generic : sig ... end

module No_percentage_toy : sig ... end

This is a toy block scanner, used in tests, and to demonstrate registration of preprocessors and diagnostics for EXEC blocks.

module Esql : sig ... end

module EXEC_MAP = Cobol_preproc.Options.EXEC_MAP

val exec_scanners : Cobol_parser.Options.exec_scanners

val more : (Cobol_parser.Options.EXEC_MAP.key * Cobol_parser.Options.exec_scanner) list -> Cobol_parser.Options.exec_scanners