Module Superbol_preprocs.No_percentage_toy

This is a toy block scanner, used in tests, and to demonstrate registration of preprocessors and diagnostics for EXEC blocks.

type Cobol_common.Exec_block.TYPES.exec_block +=

| Nopers_exec_block of Cobol_preproc.Text.t

type Cobol_common.Exec_block.TYPES.diagnostic +=

| Unexpected_character_in_exec_block of {

char : char;

loc : Cobol_common.Srcloc.TYPES.srcloc;

}

val scanner : Cobol_parser.Options.exec_scanner